FAC-513: refactor: change type of attachmentsCount to int

+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/port/in/evidence/GetAttributeEvidenceListUseCase.java
@@ -54,5 +54,5 @@ public interface GetAttributeEvidenceListUseCase {
         }
     }
 
-    record AttributeEvidenceListItem(String description, Long attachmentsCount) {}
+    record AttributeEvidenceListItem(String description, int attachmentsCount) {}
 }
